pyproject.toml
[project]
name = "program-crawler"
version = "0.2.0"
description = "AI-driven research agent for academic program information"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5,<3",
    "playwright>=1.40",
    "playwright-stealth>=2.0",
    "openpyxl>=3.1",
    "httpx>=0.25",
    "pyyaml>=6.0",
    "google-genai>=1.0",
    # Legacy (OpenAI-compatible) — kept for batch mode
    "openai>=1.10",
    "tiktoken>=0.5",
    # Optional tools
    "pyautogui>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "ruff>=0.2",
]

[project.scripts]
research = "app.runner:main"
crawl = "app.runner:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"
